The second laboratory is completed
