+++ lab_2/Crawler.py
@@ -1,24 +1,7 @@
 import os
 import csv
-import argparse
 
 from icrawler.builtin import GoogleImageCrawler
-
-import Iterator
-
-def parser_for_program() -> tuple[str, str, str]:
-    """
-    Parses the search keyword, the name of the directory
-    where the downloaded images will be saved and the name of the annotation.
-
-    :return: A list containing the keyword, the name of the directory and the name of the annotation
-    """
-    parser = argparse.ArgumentParser()
-    parser.add_argument('keyword', type=str, help='keyword')
-    parser.add_argument('save_file', type=str, help='safe_dir')
-    parser.add_argument('annotation_file', type=str, help='annotation_file')
-    args = parser.parse_args()
-    return args.keyword, args.save_file, args.annotation_file
 
 def crawler(keyword: str, max_number: int, save_dir: str) -> None:
     """
@@ -38,6 +21,7 @@ def crawler(keyword: str, max_number: int, save_dir: str) -> None:
         license='noncommercial,modify')
     google_crawler.crawl(keyword=keyword, filters=filters, max_num=max_number)
 
+
 def create_annotation(save_dir: str, annotation_path: str) -> None:
     """
     Creates an annotation in a .csv file for images
@@ -56,15 +40,3 @@ def create_annotation(save_dir: str, annotation_path: str) -> None:
             abs_path = os.path.abspath(os.path.join(save_dir, picture))
             rel_path = os.path.join(save_dir, picture)
             writer.writerow([abs_path, rel_path])
-
-def main():
-    keyword, save_dir, annotation_file = parser_for_program()
-    crawler(keyword, 100, save_dir)
-    create_annotation(save_dir,annotation_file)
-    it = iter(Iterator.IteratorForImages(annotation_file))
-
-    for image in it:
-        print("Очередное значение:", image)
-
-if __name__ == "__main__":
-     main()
